php84Extensions.soap: skip tests
See https://hydra.nixos.org/build/275613910

The issue that symbols from `session.so` cannot be loaded doesn't happen
when building a PHP with ext-soap enabled.

To me it seems as if PHP just doesn't find it on runtime. My current
hypothesis is that this is because ext-soap now depends on ext-session[1].

[1] https://github.com/php/php-src/commit/c7797fc8c0503c18791fd651f5ebdd7164b4bc48