segger-jlink: drop
This is a proprietary package that depends on the
vulnerability‐ridden Qt 4, and was removed along with Qt 4 in
2022. It was re‐added at the beginning of the year, but upstream
is unwilling to upgrade their Qt version and as it is proprietary
software we have no way to patch it on our end.

I think that it would be best not to ship this in 24.11. While it’s
clearly useful software, we made the deliberate decision to remove Qt 4
seven years after it reached end of life, and since this software is
seemingly doomed to eternal `knownVulnerabilities` I think that it does
not meet our standards for inclusion. This would be better maintained in
a third‐party repository, as suggested by the maintainer in
<https://github.com/NixOS/nixpkgs/pull/214195#discussion_r1644329222>.

This reverts commit d5fd2638269a1e140f2c30d1f3535e755183ccfb.