python3Packages.pytubefix: test failures
Two kinds of test failures show in Hydra:

1. tests failing due to lack of DNS
1. tests genuinely failing

The latter one has an open PR, which will likely fix this for an upcoming version, with the patch included here.
The first one now has an open issue upstream and which may or may not be fixed.
Disabling those tests has been suggested in #351298 already, which ended up not being merged.

Signed-off-by: benaryorg <binary@benary.org>