clang-tools: Improve wrapper script, add tests
clang-tools invokes clangd through a wrapper script whose sole purpose
is to extend `CPATH` and `CPLUS_INCLUDE_PATH` envs, so that clangd is
later able to find built-in and stdlib headers.

This script does two things wrong:

1. We extend `CPATH` et al with entries from both
`@clang@/nix-support/libc-cflags` and `@clang@/resource-root/include`,
which is kind of a self-made problem, since clangd already has some
header auto-detection logic which we (used to) circumvent (for more
details, see the diff).

2. We extend `CPATH` et al unconditionally, breaking the
`--query-driver` option. I'm not sure what's the perfect solution here,
but skipping this logic when we detect `--query-driver` should be good
enough. I mean, ideally we wouldn't be playing with `CPATH` etc.
whatsoever, but well.

Closes https://github.com/NixOS/nixpkgs/issues/351962
Closes https://github.com/NixOS/nixpkgs/issues/348791
Closes https://github.com/NixOS/nixpkgs/issues/345704