lib.attrsets: add `flattenAttrs` function
> `lib.attrsets.flattenAttrs`:
> Flatten an attribute set to the bottom name and value pairs as
> one singular final attribute set, stripping everything else
> above.
>
> In a few words, this does what `lib.lists.flatten` does, but for
> attribute sets, rather than lists.

Attribute sets are useful for structuring code, with attribute
names serving as labels, rather than relying on comments.

One example would be when working defining `environment.shellAliases.`

```nix
environment.shellAliases =
  let
    aliases = {
      fd.f = "fd --hidden --no-ignore";

      lsd = {
        ls = {
          lc = "lsd --almost-all";

          lci = "lsd --almost-all --inode";

          lac = "lsd --all";
          lca = "lsd --all";

        };
        tree = {
          tc = "lsd --almost-all --ignore-glob .git --tree";

          tcd = "lsd --almost-all --ignore-glob .git --tree --depth";

          tci = "lsd --all --ignore-glob .git --inode --tree";
          tic = "lsd --all --ignore-glob .git --inode --tree";

          tcid = "lsd --all --ignore-glob .git --inode --tree --depth";
          ticd = "lsd --all --ignore-glob .git --inode --tree --depth";
        };
      };

      ripgrep = {
        r = "rg --glob='!.git'";
        rh = "rg --glob='!.git' --hidden";
        rhn = "rg --glob='!.git' --hidden --no-ignore";
        rhnn = "rg --hidden --no-ignore";
      };
    };
  in
    flattenAttrs aliases;
```

Here, the aliases are categorized by the binary that is being aliased.
The `lsd` set also has another subsection layer for `ls` and `tree`.
Comments could be used, but an in-code approach is better, especially
when the attribute set is large and not just a brief example like
what is shown here. This allows for inspection using
Nix directly along with any other transformations that are desired.

`environment.shellAliases` only takes one attribute set with direct
string name-value pairs. `flattenAttrs` will strip the "labels" and
return only the bottom level across the passed attribute set, being
equivalent to:

```nix
environment.shellAliases = {
  # fd
  f = "fd --hidden --no-ignore";

  # lsd - ls
  lc = "lsd --almost-all";

  lci = "lsd --almost-all --inode";

  lac = "lsd --all";
  lca = "lsd --all";

  # lsd - tree
  tc = "lsd --almost-all --ignore-glob .git --tree";

  tcd = "lsd --almost-all --ignore-glob .git --tree --depth";

  tci = "lsd --all --ignore-glob .git --inode --tree";
  tic = "lsd --all --ignore-glob .git --inode --tree";

  tcid = "lsd --all --ignore-glob .git --inode --tree --depth";
  ticd = "lsd --all --ignore-glob .git --inode --tree --depth";

  # ripgrep
  r = "rg --glob='!.git'";
  rh = "rg --glob='!.git' --hidden";
  rhn = "rg --glob='!.git' --hidden --no-ignore";
  rhnn = "rg --hidden --no-ignore";
};
```

This brings the usefulness of `lib.lists.flatten` to attribute
sets. This doesn't conflict with nor modify any existing
implementations, this is an addition, so this change should be safe.