ffmpeg: enable placebo
/nix/store/0g381zbwph4i1xwr3r0jjgl3ld22ay9y-ffmpeg-7.1-bin	 771250848
/nix/store/p9pnzdynjr2g6rxgrbinljsxanzc5vh5-ffmpeg-7.1-bin	 795740304

With a bigger closure containing gtk4 which would be quite basic for any headed
setup:

    nix-build --expr 'with import ./. { }; buildEnv { name = "foo"; paths = [ ffmpeg gtk4 ]; }'

/nix/store/br81b18qsqa9pyd6xm03x1dh2h7mgs4r-foo	 967359472
/nix/store/f34hzrcm58y1kzrkc1bkwrip79f2yrc6-foo	 991848928

That's still the same kind of increase and it stems from shaderc and libdovi,
both of which aren't used in typical desktop packages.

The only users who would be unaffected by this closure size increase are those
who already use libplacebo through the popular video player `mpv`.