cc-wrapper: Add nixSupport.cc-cflags-link
Prior to this patch, I could not find a way to do `stdenv.cc.override {
nixSupport.* = ...; }` that would enable passing flags to the linker
driver.

There is `nixSupport.cc-ldflags`, but those flags go directly to the
linker via NIX_LDFLAGS which uses `-Wl,$flag`, and some flags need to be
sent to the driver via NIX_CFLAGS_LINK.

There are some flags which want to only be passed to the linker,
otherwise they emit compiler driver warnings about unused arguments,
which can be noisy.

Signed-off-by: Peter Waller <p@pwaller.net>