cc-wrapper: Add nixSupport.cc-cflags-link
Prior to this patch, I could not find a way to do `stdenv.cc.override {
nixSupport.* = ...; }` that would enable passing flags such as
`-fsanitize=address` to the linker driver.

There is `nixSupport.cc-ldflags`, but those flags go directly to the
linker via NIX_LDFLAGS which uses `-Wl,$flag`, and some flags need to be
sent to the driver via NIX_CFLAGS_LINK.

Another possibility I considered is to pass nixSupport.cc-cflags to the
link invocation, but I understand this is not considered ideal, leading
to unused flag warnings at best, and unintended behaviour at worst. It
is probably better to be able to explicitly pass flags to the linker
invocation.

Signed-off-by: Peter Waller <p@pwaller.net>