Automate switching to new GE-Proton version after update
The purpose of this package is to automate the downloading/updating of GE-Proton. Since steam only sees the most current version, we have to then switch to the new version manually in the steam UI.

Inspired by https://aur.archlinux.org/packages/proton-ge-custom-bin, this changes the version as steam sees it to always just be "GE-Proton", so it doesn't have to be changed and always points to the most recent version.