nixos/zigbee2mqtt: only add port to DeviceAllow if it is a path
zigbee2mqtt supports having non-device ports (e.g. `tcp://`); those
should not be set in DeviceAllow. No URI will start with `/`, so use
that as the filter that it is a “real” device that needs to be allowed.

(cherry picked from commit 577e162073db726c827d567141f0e214d8cfd913)