zxing-cpp: c++20 for char8_t
build fails on clang-19 due to using c++ 17 and having char8_t typedefed
to an uint8_t, which is not supported by std::char_traits. Using c++20
fixes this as the code will use a char8_t.

Used a substituteInPlace so the override can be removed once c++20 is
the default in the cmake file.