folly: fix split outputs
Downstream derivations were picking up references to the `dev` output
through indirect `__FILE__` usage in the headers. I borrowed and
modified this trick from the Boost package, where I believe vcunat
originated it. It’s not very pretty and it should be in a hook or
something, but it works at a pinch.