nixos-option: clean up includes based on clang-tidy suggestions
Before the code was using include-what-you-means, but it's a bit tricky
to integrate this in nix. I had some trouble with nix-eval-jobs when using include-what-you-means.
clang-tidy on the other hand proved more robust to me and is useful for linting in general,
so less tooling overall for this little project.