lib.platform: optimise platform constraint algebra
This originally had some more opinionated changes and was based on an older
state of my patches. I only kept the performance optimisations.

I was not able to measure a significant change in perf but these look like
they'd be better for perf to me and also improve the code.

Co-authored-by: Atemu <git@atemu.net>