nixos/doc/rl-2505: add more comments about ordering
This looks a little ridiculous right now, but my experience is that
it’s common to find the beginning or end of a section and add more
things there without seeing the comments. We should probably move
to a one file per release note system, but in the meantime this is
a low‐cost way to help reduce merge conflicts.