stdenv/linux: apply default compiler errors from GCC 14
Let’s make sure both OSes get an adequate amount of compiler fun this
cycle! This is a version of alyssais’s
<https://github.com/NixOS/nixpkgs/pull/344239> that applies to the Linux
standard environment’s compiler wrapping rather than the versioned GCC
packages themselves.

This is purely a transition mechanism to help us prepare for the GCC
14 update while GCC 14 itself is not yet ready for wide deployment;
it will be reverted before 25.05, whether because it becomes obsolete
thanks to an upgrade to GCC 14 or because we give up on GCC 14 and
deicde to stick with GCC 13 with the vanilla flags.

This approach means that the standard `gcc` will be wrapped correctly,
unlike with the previous PR, achieving the primary goal of the change,
but `gcc13` will remain as it was before. Currently, I believe
the difference between the two is just bootstrapping stage stuff;
it’s a bit weird to me to diverge, especially since there is a
definition of `gcc` in `all-packages.nix` that looks like it ought
to make them equivalent, but it may have the nice property that we
can pin especially stubborn packages to GCC 13 without them getting
the errors (though I am not sure if this is actually true or not,
since I’m not sure how `gcc13Stdenv` actually works).

One downside is that this only applies to Linux, but since this is
essentially meant to be a forcing function to get us to deal with
the errors early, that seems fine.

Source: https://gcc.gnu.org/gcc-14/porting_to.html
Co-authored-by: Alyssa Ross <hi@alyssa.is>