stdenv.mkDerivation: refactor output checks
This just moves the code around a little bit to be able to build on it
in follow up commits without too much of a diff.

Adding to removedOrReplacedAttrNames is just a cleanup right now and
doesn't change anything functionally, yet. This will be required for
later on to avoid having structured outputChecks and those directly on
the derivation at the same time.