chromium: remove "channel" argument
This is no longer used as we only differentiate between stable and
ungoogled-chromium, which we already track in the "ungoogled" boolean.

Beta and dev channels are gone for good.
It's been a year since their removal in 59719f787e94f39e64e9086d08eaedd8a9e61b22.

There is, however, an additional channel mapping in nixos/tests/chromium
but that one is independent from this one here.