jujutsu: skip a flaky test
A user reported that there was nothing available in the binary cache for
`aarch64-linux` for the latest release, and some investigation turned up that
the runners ran into a flaky test that was fixed recently.

For simplicity (and so that we don't have to backport the upstream patch), skip
this test. The feature works perfectly fine, it is only the test itself that
is buggy.

This should be removed next month after the 0.24.0 release.

Signed-off-by: Austin Seipp <aseipp@pobox.com>