nv-codec-headers: multiple-version support
This is a proof-of-concept for a multiple-version package.

Since nv-codec-headers is a very simple package - it just throws files into the
final destination -, it is easily parameterizable.

After some experimentation, I decided to put all the packages inside
`passthru.versions`. This way any of them can be reached by calling
`nv-codec-headers.passthru.v<major version>`.