mpv: restore non‐obsolete Swift library path hack
Turns out that it’s not obsolete; it avoids a runtime dependency
on the Swift wrapper because of some weird stuff we do. This should
probably be fixed in the Swift packaging, but let’s fix the closure
size again for now and add a comment for that.

This reverts commit 84d64ade98bd69f4dbe3534560e15dae2839d1d4.

Fixes: #356861
(cherry picked from commit 8d9ac2131da3f6938512912b874d0ff22cb1a48a)