skawarePackages: split man page packages back out
These man pages are maintained by a different person as separate
packages, and aren’t necessarily released in lockstep with the
software they document, meaning they can be out of date and it
doesn’t make sense to couple them directly.

This change brings the build‐time closure of `execline` down from
426 packages to 367, which is a meaningful difference considering
it’s used as the default emulator for host platforms that can be
executed directly by a build platform.

We could include these as `passthru.man` for partial compatibility,
but it would still have the version mismatch issue, and wouldn’t
retain the same behaviour in all contexts, so I opted not to.

This reverts about half of commit
7cadc175919016d329c868915f1f1c42fc8fb817.