flutter: revert remove usages of aliases {build,host,target}Platform
PR #350299 broke Flutter's update script with how it handled getting the
platform. Furthermore, the problem was made worse in the last update to
Flutter which had to work around that problem. We'll just revert it for
now and do a proper fix once we can figure out how to satify setting the
different platforms without causing any more problems.