replaceVars: allow exemptions
replaceVars has a checkPhase to confirm that no left-over @...@ patterns
remain. Since it's not possible to use "placeholder" with replaceVars,
the substitution of some patterns must be delayed to a later step.

By passing "null" for those keys explicitly, replaceVars can make an
exemption *just for this case*, but keep checking all other references.