qmk: use the default GCC version for AVR
Upstream previously recommended GCC 8 when targeting AVR because
of binary size concerns. They have since removed this warning in
<https://github.com/qmk/qmk_firmware/pull/24206>, as more recent GCC
versions have resolved the issues.

Building `keyboardio/model01` before:

     * The firmware size is fine - 26984/28672 (94%, 1688 bytes free)

And after:

     * The firmware size is fine - 27200/28672 (94%, 1472 bytes free)

+216 bytes (1.008× the total size) is not a big deal. Since upstream
is fine with it these days, we can be too, and not keep around an
unsupported compiler just for QMK.